requirements.txt
# Core packages
streamlit>=1.34.0
pandas>=2.2.0
numpy<2.0
matplotlib>=3.8.0
plotly>=5.20.0

# PDF support
PyPDF2>=3.0.1

# LangChain + OpenAI
openai>=1.30.0
langchain>=0.1.17
langchain-openai>=0.1.6

# For environment variable management
python-dotenv>=1.0.1

# Optional utils
requests>=2.31.0
pyarrow>=15.0.0
scikit-learn>=1.4.0
openpyxl>=3.1.0

# Fix for NumPy 2.x compatibility issues
pybind11>=2.12
bottleneck>=1.3.7
numexpr>=2.9.0
